few memory updates

+++ sugarloaf/src/layout/glyph.rs
@@ -15,11 +15,11 @@ pub const GLYPH_DETAILED: u32 = 0x80000000;
 #[derive(Copy, Debug, Clone)]
 pub struct GlyphData {
     pub data: u32,
-    pub size: usize,
+    pub size: u32,
 }
 
 impl GlyphData {
-    pub fn simple(id: u16, advance: f32, size: usize) -> Self {
+    pub fn simple(id: u16, advance: f32, size: u32) -> Self {
         let advance = (advance * 64.).max(0.) as u32;
         Self {
             data: id as u32 | ((advance & 0x7FFF) << 16),
